Basic locking functionality for crates and playlists.

+++ mixxx/src/library/cratetablemodel.cpp
@@ -117,29 +117,38 @@ QString CrateTableModel::getTrackLocation(const QModelIndex& index) const {
 }
 
 void CrateTableModel::removeTracks(const QModelIndexList& indices) {
-    const int trackIdIndex = fieldIndex(LIBRARYTABLE_ID);
-
-    QList<int> trackIds;
-    foreach (QModelIndex index, indices) {
-        int trackId = index.sibling(index.row(), fieldIndex(LIBRARYTABLE_ID)).data().toInt();
-        trackIds.append(trackId);
-    }
-
     CrateDAO& crateDao = m_pTrackCollection->getCrateDAO();
-    foreach (int trackId, trackIds) {
-        crateDao.removeTrackFromCrate(trackId, m_iCrateId);
-    }
+    bool locked = crateDao.isCrateLocked(m_iCrateId);
 
-    select();
+    if (!locked) {
+        const int trackIdIndex = fieldIndex(LIBRARYTABLE_ID);
+
+        QList<int> trackIds;
+        foreach (QModelIndex index, indices) {
+            int trackId = index.sibling(index.row(), fieldIndex(LIBRARYTABLE_ID)).data().toInt();
+            trackIds.append(trackId);
+        }
+
+        foreach (int trackId, trackIds) {
+            crateDao.removeTrackFromCrate(trackId, m_iCrateId);
+        }
+
+        select();
+    }
 }
 
 void CrateTableModel::removeTrack(const QModelIndex& index) {
-    const int trackIdIndex = fieldIndex(LIBRARYTABLE_ID);
-    int trackId = index.sibling(index.row(), trackIdIndex).data().toInt();
-    if (m_pTrackCollection->getCrateDAO().removeTrackFromCrate(trackId, m_iCrateId)) {
-        select();
-    } else {
-        // TODO(XXX) feedback
+    CrateDAO& crateDao = m_pTrackCollection->getCrateDAO();
+    bool locked = crateDao.isCrateLocked(m_iCrateId);
+
+    if (!locked) {
+        const int trackIdIndex = fieldIndex(LIBRARYTABLE_ID);
+        int trackId = index.sibling(index.row(), trackIdIndex).data().toInt();
+        if (m_pTrackCollection->getCrateDAO().removeTrackFromCrate(trackId, m_iCrateId)) {
+            select();
+        } else {
+            // TODO(XXX) feedback
+        }
     }
 }
